+++ src/cqspoolterrain/__init__.py
@@ -1,11 +1,3 @@
-from .powerLine import (
-    power_line_straight, 
-    power_line_curve, 
-    make_power_face, 
-    make_curved_connectors, 
-    make_curved_magnets,
-    power_line_end
-)
 from .Base import Base
 from .Spool import Spool
 from .Cradle import Cradle