+++ src/cqspoolterrain/StairLift.py
@@ -18,6 +18,7 @@ from . import Base
 from cqterrain import tile
 from cadqueryhelper import wave
 
+
 class StairLift(Base):
     def __init__(
             self,