+++ src/cqspoolterrain/__init__.py
@@ -1,4 +1,5 @@
 from .Base import Base
+from .industrialStairs import industrial_stairs
 from .Spool import Spool
 from .Cradle import Cradle
 from .StairLift import StairLift
@@ -6,5 +7,6 @@ from .ControlPlatform import ControlPlatform
 from .SpoolCladding import SpoolCladding
 from .SpoolCladdingGreebled import SpoolCladdingGreebled
 from .SpoolCladdingGreebledUnique import SpoolCladdingGreebledUnique
+
 
 from .PowerStation import PowerStation