+++ src/cqspoolterrain/pipe/__init__.py
@@ -4,3 +4,4 @@ from .curve import curve
 from .end import end
 from .connector import connector
 from .hatch import hatch
+from .platform import platform